fix: Move enums to subfolders by regenerating
