+++ tests/serverpod_test_client/lib/src/protocol/protocol.dart
@@ -16,9 +16,9 @@ import 'object_with_enum.dart' as _i6;
 import 'object_with_maps.dart' as _i7;
 import 'object_with_object.dart' as _i8;
 import 'serverOnly/default_server_only_class.dart' as _i9;
-import 'default_server_only_enum.dart' as _i10;
+import 'serverOnly/default_server_only_enum.dart' as _i10;
 import 'serverOnly/not_server_only_class.dart' as _i11;
-import 'not_server_only_enum.dart' as _i12;
+import 'serverOnly/not_server_only_enum.dart' as _i12;
 import 'simple_data.dart' as _i13;
 import 'simple_data_list.dart' as _i14;
 import 'test_enum.dart' as _i15;
@@ -39,9 +39,9 @@ export 'object_with_enum.dart';
 export 'object_with_maps.dart';
 export 'object_with_object.dart';
 export 'serverOnly/default_server_only_class.dart';
-export 'default_server_only_enum.dart';
+export 'serverOnly/default_server_only_enum.dart';
 export 'serverOnly/not_server_only_class.dart';
-export 'not_server_only_enum.dart';
+export 'serverOnly/not_server_only_enum.dart';
 export 'simple_data.dart';
 export 'simple_data_list.dart';
 export 'test_enum.dart';
