+++ tests/serverpod_test_client/lib/src/protocol/serverOnly/default_server_only_enum.dart
