+++ tests/serverpod_test_client/lib/src/protocol/serverOnly/not_server_only_enum.dart
